+++ src/components/Verse/GroupedVerseAndTranslation.tsx
@@ -1,25 +1,14 @@
 /* eslint-disable unicorn/no-array-reduce */
 import { useMemo } from 'react';
 
-import useTranslation from 'next-translate/useTranslation';
-import { shallowEqual, useSelector } from 'react-redux';
-import useSWR from 'swr/immutable';
-
 import styles from './GroupedVerseAndTranslation.module.scss';
 import PlainVerseText from './PlainVerseText';
 
 import Error from '@/components/Error';
 import TranslationText from '@/components/QuranReader/TranslationView/TranslationText';
 import Spinner from '@/dls/Spinner/Spinner';
-import useQcfFont from '@/hooks/useQcfFont';
-import { selectQuranReaderStyles } from '@/redux/slices/QuranReader/styles';
-import { selectSelectedTranslations } from '@/redux/slices/QuranReader/translations';
-import { getDefaultWordFields, getMushafId } from '@/utils/api';
-import { makeVersesUrl } from '@/utils/apiPaths';
-import { areArraysEqual } from '@/utils/array';
+import useVerseAndTranslation from '@/hooks/useVerseAndTranslation';
 import { getVerseWords } from '@/utils/verse';
-import { fetcher } from 'src/api';
-import { VersesResponse } from 'types/ApiResponses';
 
 interface Props {
   chapter: number;
@@ -27,34 +16,9 @@ interface Props {
   to: number;
 }
 
-const GroupedVerseAndTranslation: React.FC<Props> = ({ chapter, from, to }) => {
-  const { lang } = useTranslation();
-  const translations = useSelector(selectSelectedTranslations, areArraysEqual);
-  const { quranFont, mushafLines, translationFontScale } = useSelector(
-    selectQuranReaderStyles,
-    shallowEqual,
-  );
+const GroupedVerseAndTranslation: React.FC<Props> = (props) => {
+  const { data, error, mutate, translations, translationFontScale } = useVerseAndTranslation(props);
 
-  const mushafId = getMushafId(quranFont, mushafLines).mushaf;
-  const apiParams = {
-    ...getDefaultWordFields(quranFont),
-    translationFields: 'resource_name,language_id',
-    translations: translations.join(','),
-    mushaf: mushafId,
-    from: `${chapter}:${from}`,
-    to: `${chapter}:${to}`,
-  };
-
-  const shouldFetchData = !!from;
-
-  const { data, error, mutate } = useSWR<VersesResponse>(
-    shouldFetchData ? makeVersesUrl(chapter, lang, apiParams) : null,
-    fetcher,
-  );
-
-  useQcfFont(quranFont, data?.verses ? data.verses : []);
-
-  // Move useMemo before conditionals
   const allWords = useMemo(() => {
     if (!data?.verses) return [];
     return data.verses.reduce((acc, verse) => {
@@ -62,11 +26,9 @@ const GroupedVerseAndTranslation: React.FC<Props> = ({ chapter, from, to }) => {
     }, []);
   }, [data?.verses]);
 
-  // Add memoized grouped translations
   const groupedTranslations = useMemo(() => {
     if (!data?.verses) return {};
     return translations.reduce<Record<number, any>>((acc, translationId: number) => {
-      // eslint-disable-next-line unicorn/no-array-reduce
       const texts = data.verses.reduce((textsAcc, verse) => {
         const translation = verse.translations?.find(
           (t) => String(t.resourceId) === String(translationId),
@@ -83,30 +45,38 @@ const GroupedVerseAndTranslation: React.FC<Props> = ({ chapter, from, to }) => {
 
   return (
     <div className={styles.container}>
-      <PlainVerseText words={allWords} />
-      {/* Translations section */}
-      {translations.map((translationId: number) => {
-        const translationGroup = groupedTranslations[translationId];
-        if (!translationGroup?.length) return null;
+      <div className={styles.arabicSection}>
+        <div className={styles.arabicVerseContainer}>
+          <PlainVerseText words={allWords} />
+        </div>
+      </div>
 
-        const firstTranslation = translationGroup[0];
-        return (
-          <div key={`translation-${translationId}`}>
-            <TranslationText
-              languageId={firstTranslation.languageId}
-              resourceName={firstTranslation.resourceName}
-              translationFontScale={translationFontScale}
-              text={translationGroup
-                .map((t, index) =>
-                  data?.verses?.length > 1
-                    ? `${t.text} (${data.verses[index].verseNumber})`
-                    : t.text,
-                )
-                .join(' ')}
-            />
-          </div>
-        );
-      })}
+      <div className={styles.translationsSection}>
+        {translations.map((translationId: number) => {
+          const translationGroup = groupedTranslations[translationId];
+          if (!translationGroup?.length) return null;
+
+          const firstTranslation = translationGroup[0];
+          return (
+            <div key={`translation-${translationId}`} className={styles.translationGroup}>
+              <div className={styles.translationContainer}>
+                <TranslationText
+                  languageId={firstTranslation.languageId}
+                  resourceName={firstTranslation.resourceName}
+                  translationFontScale={translationFontScale}
+                  text={translationGroup
+                    .map((t, index) =>
+                      data.verses.length > 1
+                        ? `${t.text} (${data.verses[index].verseNumber})`
+                        : t.text,
+                    )
+                    .join(' ')}
+                />
+              </div>
+            </div>
+          );
+        })}
+      </div>
     </div>
   );
 };
