+++ src/components/Verse/VerseAndTranslation.tsx
@@ -1,23 +1,11 @@
-import useTranslation from 'next-translate/useTranslation';
-import { shallowEqual, useSelector } from 'react-redux';
-import useSWR from 'swr/immutable';
-
-import Spinner from '../dls/Spinner/Spinner';
-import TranslationText from '../QuranReader/TranslationView/TranslationText';
-
 import PlainVerseText from './PlainVerseText';
 import styles from './VerseAndTranslation.module.scss';
 
 import Error from '@/components/Error';
-import useQcfFont from '@/hooks/useQcfFont';
-import { selectQuranReaderStyles } from '@/redux/slices/QuranReader/styles';
-import { selectSelectedTranslations } from '@/redux/slices/QuranReader/translations';
-import { getDefaultWordFields, getMushafId } from '@/utils/api';
-import { makeVersesUrl } from '@/utils/apiPaths';
-import { areArraysEqual } from '@/utils/array';
+import TranslationText from '@/components/QuranReader/TranslationView/TranslationText';
+import Spinner from '@/dls/Spinner/Spinner';
+import useVerseAndTranslation from '@/hooks/useVerseAndTranslation';
 import { getVerseWords } from '@/utils/verse';
-import { fetcher } from 'src/api';
-import { VersesResponse } from 'types/ApiResponses';
 
 /**
  * Given a verse range
@@ -34,32 +22,8 @@ interface Props {
   to: number;
 }
 
-const VerseAndTranslation: React.FC<Props> = ({ chapter, from, to }) => {
-  const { lang } = useTranslation();
-  const translations = useSelector(selectSelectedTranslations, areArraysEqual);
-  const { quranFont, mushafLines, translationFontScale } = useSelector(
-    selectQuranReaderStyles,
-    shallowEqual,
-  );
-
-  const mushafId = getMushafId(quranFont, mushafLines).mushaf;
-  const apiParams = {
-    ...getDefaultWordFields(quranFont),
-    translationFields: 'resource_name,language_id',
-    translations: translations.join(','),
-    mushaf: mushafId,
-    from: `${chapter}:${from}`,
-    to: `${chapter}:${to}`,
-  };
-
-  const shouldFetchData = !!from;
-
-  const { data, error, mutate } = useSWR<VersesResponse>(
-    shouldFetchData ? makeVersesUrl(chapter, lang, apiParams) : null,
-    fetcher,
-  );
-
-  useQcfFont(quranFont, data?.verses ? data.verses : []);
+const VerseAndTranslation: React.FC<Props> = (props) => {
+  const { data, error, mutate, translationFontScale } = useVerseAndTranslation(props);
 
   if (error) return <Error error={error} onRetryClicked={mutate} />;
